requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
cdifflib==1.2.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.0
chumpy==0.70
click==8.1.7
cmake==3.27.7
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
dnspython==2.6.1
editdistance==0.8.1
einops==0.8.0
einx==0.3.0
email_validator==2.2.0
encodec==0.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi-cli==0.0.4
fastjsonschema==2.19.1
# Editable install with no version control (featup==0.1.2)
-e /data_sdf/lgattention-train/FeatUp
ffmpy==0.3.2
filelock==3.12.4
Flask==3.0.2
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
google==3.0.0
gradio==4.36.1
gradio_client==1.0.1
grpcio==1.62.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
hydra-core==0.11.3
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.4.0
inflect==7.3.0
iopath==0.1.10
ipython==8.25.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
KNN-CUDA @ https://github.com/unlimblue/KNN_CUDA/releases/download/0.2/KNN_CUDA-0.2-py3-none-any.whl#sha256=809a43da2eed950c65ba15e355e8cdbf975b6a948dced1f02414f26aba586cd0
kornia==0.7.3
kornia_rs==0.1.5
lightning-utilities==0.11.6
lit==17.0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbformat==5.9.2
nemo_text_processing==1.0.2
nest-asyncio==1.6.0
networkx==3.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.9.0.80
orjson==3.10.5
packaging==23.2
pandas==2.1.1
parso==0.8.4
peft==0.5.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.19.0
plyfile==1.0.3
pointnet2==3.0.0
portalocker==2.8.2
potpourri3d==1.0.0
prompt_toolkit==3.0.47
protobuf==5.26.0
psbody-mesh==0.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pynini==2.1.5
PyOpenGL==3.1.7
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==0.7.1
-e git+https://github.com/facebookresearch/pytorch3d.git@f34104cf6ebefacd7b7e07955ee7aaa823e616ac#egg=pytorch3d
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.10.3
requests==2.32.3
retrying==1.3.4
rich==13.7.1
robust-laplacian==0.2.7
rouge==1.0.1
rpds-py==0.18.0
ruff==0.4.10
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.3.0
timm==0.4.12
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0
torch-geometric==2.6.1
torch-scatter==2.1.2
torchaudio==2.2.0
torchmetrics==1.4.1
torchvision==0.17.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.39.1
trimesh==4.1.4
triton==2.2.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
vector-quantize-pytorch==1.14.30
vocos==0.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.1
WeTextProcessing==1.0.2
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
zmq==0.0.0
